fix bug url production

+++ database/migrations/2025_01_23_201441_create_wisata_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('wisata', function (Blueprint $table) {
             $table->increments('id_wisata');
-            $table->unsignedBigInteger('id_mobil');
+            $table->unsignedInteger('id_mobil');
             $table->string('nama');
             $table->string('harga');
             $table->string('wilayah');
